AfyaMum-Backend/AfyaMum/pyproject.toml
[tool.poetry]
name = "afyamum"
version = "0.1.0"
description = ""
authors = ["ingenious developer <72666669+ingenious-dev@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
django = "^4.2.4"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.2.0"
djangorestframework-simplejwt = "^5.2.2"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
